+++ pis/__init__.py
@@ -1,3 +1,3 @@
 """pis — Python Install Shit. A tiny pip alternative."""
 
-__version__ = "0.0.5"
+__version__ = "0.0.6"